mem-cache: Annotate unused variables and cast size conversions in compressed tags and super block

+++ src/mem/cache/tags/compressed_tags.test.cc
@@ -70,7 +70,7 @@ class SuperBlkTestFixture : public ::testing::Test
     verifyInvariants(const SuperBlk &sb)
     {
         uint8_t count_valid = 0;
-        uint8_t min_cf = sb.blks.size();
+        uint8_t min_cf = static_cast<uint8_t>(sb.blks.size());
         std::size_t total_bits = 0;
         for (const auto &blk : sb.blks) {
             if (blk->isValid()) {
